requirements.txt
absl-py==2.2.1
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Django==5.1.6
django-cors-headers==4.7.0
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pywin32==308
pyzmq==26.2.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
